ServiceDiscovery shortnames working
We're now adding an environmental variable $SEARCH_DOMAIN to TaskDefinitions.
This is just a service discovery namespace that looks like a domain name.

Containers that want to connect to other services must start with this bash
snippet.

if [ $SEARCH_DOMAIN ]; then echo "search ${SEARCH_DOMAIN}" >> /etc/resolv.conf; fi

This will make the running container automatically look in the service discovery
namespace when it attempts host lookups.

Whatever code is being run in the container then, do not need to worry about
dealing with some FQDN. It can just connect to a service like it's connecting to
a host. We can deploy this to multiple environments without having to change any
code at all.